# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Recyclify Bidder Market CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pipeline/boards:
    get:
      operationId: listPipelineBoards
      tags: [pipeline]
      responses:
        "200":
          description: List of pipeline boards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineBoardListResponse"
    post:
      operationId: createPipelineBoard
      tags: [pipeline]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PipelineBoardInput"
      responses:
        "201":
          description: Board created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineBoard"
  /pipeline/boards/{id}:
    patch:
      operationId: updatePipelineBoard
      tags: [pipeline]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PipelineBoardInput"
      responses:
        "200":
          description: Board updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineBoard"
    delete:
      operationId: deletePipelineBoard
      tags: [pipeline]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /pipeline/boards/{boardId}/stages:
    get:
      operationId: listPipelineStages
      tags: [pipeline]
      parameters:
        - name: boardId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stages for a board
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineStageListResponse"
    post:
      operationId: createPipelineStage
      tags: [pipeline]
      parameters:
        - name: boardId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PipelineStageInput"
      responses:
        "201":
          description: Stage created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineStage"
  /pipeline/stages/{id}:
    patch:
      operationId: updatePipelineStage
      tags: [pipeline]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PipelineStageInput"
      responses:
        "200":
          description: Stage updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineStage"
    delete:
      operationId: deletePipelineStage
      tags: [pipeline]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── CONTACTS ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    PipelineBoard:
      type: object
      required: [id, name, color, isDefault, pipelineType, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        isDefault:
          type: boolean
        pipelineType:
          type: string
        createdById:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    PipelineBoardInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        color:
          type: string
        pipelineType:
          type: string
    PipelineBoardListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/PipelineBoard"
    PipelineStage:
      type: object
      required: [id, boardId, name, color, position, createdAt, updatedAt]
      properties:
        id:
          type: integer
        boardId:
          type: integer
        name:
          type: string
        color:
          type: string
        position:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    PipelineStageInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        color:
          type: string
        position:
          type: integer
    PipelineStageListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/PipelineStage"
...