# lib/api-zod/src/generated/types/pipelineBoard.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Recyclify Bidder Market CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface PipelineBoard {
  id: number;
  name: string;
  color: string;
  isDefault: boolean;
  pipelineType: string;
  /** @nullable */
  createdById?: number | null;
  createdAt: string;
  updatedAt: string;
}