# lib/api-zod/src/generated/types/pipelineBoardInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Recyclify Bidder Market CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface PipelineBoardInput {
  name: string;
  color?: string;
  pipelineType?: string;
}